changes to remove dependency on router while booting up msal service
