+++ lib/msal-angular/src/index.ts
@@ -5,3 +5,4 @@ export { MsalService } from './msal.service';
 export { MsalGuard } from './msal-guard.service';
 export { MsalInterceptor } from './msal.interceptor';
 export { MsalModule } from './msal.module';
+export { MsalRouter } from './msal.router';
